auth/requirements.txt
Flask==1.1.4
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-Security==3.0.0
flask-restplus==0.13.0
Flask-Migrate==3.1.0
psycopg2-binary==2.9.3
pydantic==1.8.2
flask_redis==0.4.0
gevent==21.12.0
typer==0.4.0
Werkzeug==0.16.1
email-validator==1.1.3
python-dotenv==0.19.2
pytest==6.1.2
markupsafe==2.0.1
requests==2.27.1
Authlib==0.15.5
jaeger-client==4.8.0
Flask-OpenTracing==1.1.0
user-agents==2.2.0
